feat: open in fullscreen
